Put polygon objects in package polygonobjects
In order to divide the code into different files for better overview and
easier testing/running I created the package 'polygonobjects' and moved
the polygon object classes in there.
